fuzzydict.py
class FuzzyDict(dict):
    def __init__(self, threshold):
        self.threshold = float(threshold)
    def _get_fuzzy_elements(self, key):
        if self.has_key(key):
            yield key, self[key]
        chars = int(len(key) * self.threshold)
        keys = [x for x in self.keys() if len(x) == len(key) and x != key]
        for k in keys:
            match = 0
            for i in range(0, len(key), 1):
                if key[i] == k[i]:
                    match += 1
            if chars <= match and self.has_key(k):
                yield k, self[k]
        StopIteration()
    def assign_object(self, obj):
        if isinstance(obj, dict):
            for k, v in obj.items():
                self[k] = v
        else:
            raise TypeError()
    def fuzzy_keys(self, key):
        for k, v in self._get_fuzzy_elements(key):
            yield k
        StopIteration()
    def fuzzy_values(self, key):
        for k, v in self._get_fuzzy_elements(key):
            yield v
        StopIteration()
    def fuzzy_items(self, key):
        for k, v in self._get_fuzzy_elements(key):
            yield k, v
        StopIteration()
    def fuzzy_assign(self, key, value):
        for k, v in self._get_fuzzy_elements(key):
            self[k] = value
    def fuzzy_add(self, key, value):
        for k, v in self._get_fuzzy_elements(key):
            self[k] += value
